Remove AniTopAppBar
