--- app/shared/src/commonMain/kotlin/ui/foundation/Placeholders.kt
@@ -1,2 +1,0 @@
-package me.him188.ani.app.ui.foundation
-
