+++ app/shared/src/commonMain/kotlin/ui/foundation/widgets/TopAppBarGoBackButton.kt
@@ -1,0 +1,46 @@
+package me.him188.ani.app.ui.foundation.widgets
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.outlined.ArrowBack
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import me.him188.ani.app.navigation.LocalBackHandler
+
+
+@Composable
+fun TopAppBarGoBackButton() {
+    val handler = LocalBackHandler.current
+    TopAppBarGoBackButton {
+        handler.onBackPress()
+    }
+}
+
+@Composable
+fun TopAppBarGoBackButton(goBack: () -> Unit) {
+    TopAppBarActionButton(goBack) {
+        Icon(
+            Icons.AutoMirrored.Outlined.ArrowBack,
+            null,
+        )
+    }
+}
+
+@Composable
+fun TopAppBarActionButton(
+    onClick: () -> Unit,
+    content: @Composable () -> Unit,
+) {
+    IconButton(
+        onClick,
+//        Modifier.offset(x = (-8).dp, y = (-8).dp).width(36.dp + 16.dp).height(36.dp + 16.dp)
+    ) { // 让可点击区域大一点, 更方便
+        Box(Modifier.size(24.dp)) {
+            content()
+        }
+    }
+}
